requirements.txt
traitlets
Adafruit_MotorHat
Adafruit-SSD1306
inputs
pi-ina219